+++ tests/tt_metal/tt_metal/api/tensor/test_transfer_to_device.cpp
@@ -2,18 +2,13 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-// Tests for TransferToDevice API - transferring data from HostTensor to DeviceTensor.
+// Tests for TransferToDevice/TransferToHost API - round-trip data transfer between HostTensor and DeviceTensor.
 
 #include <gtest/gtest.h>
 #include <numeric>
 #include <vector>
 
-#include <tt-metalium/host_api.hpp>
-#include <tt-metalium/distributed.hpp>
-#include <tt-metalium/mesh_device.hpp>
 #include <tt-metalium/mesh_command_queue.hpp>
-#include <tt-metalium/mesh_buffer.hpp>
-#include <tt-metalium/host_buffer.hpp>
 
 #include <tt-metalium/experimental/tensor/host_tensor.hpp>
 #include <tt-metalium/experimental/tensor/device_tensor.hpp>
@@ -28,10 +23,33 @@ namespace {
 
 using namespace tt::tt_metal;
 
-// Use GenericMeshDeviceFixture for proper multi-device handling (same as test_tensor_topology.cpp)
-using TransferToDeviceTest = GenericMeshDeviceFixture;
+// Single device (1x1 mesh) fixture for testing tensor APIs
+class SingleDeviceMeshFixture : public MeshDeviceFixtureBase {
+protected:
+    SingleDeviceMeshFixture() :
+        MeshDeviceFixtureBase(Config{.mesh_shape = distributed::MeshShape{1, 1}, .num_cqs = 1}) {}
+};
 
-TEST_F(TransferToDeviceTest, BasicTransfer) {
+using TransferToDeviceTest = SingleDeviceMeshFixture;
+
+// TODO: This should be a public utlity:
+// Helper function to create a HostTensor for receiving data from device.
+// Creates a HostTensor with pre-allocated float buffer using from_vector.
+HostTensor create_receiving_host_tensor(const TensorSpec& tensor_spec) {
+    const size_t num_elements = tensor_spec.logical_shape().volume();
+    std::vector<float> data(num_elements, 0.0f);
+    return HostTensor::from_vector(std::move(data), tensor_spec);
+}
+
+// Helper function to extract float data from a HostTensor's buffer directly.
+// This works for ROW_MAJOR layout where logical shape matches physical shape.
+std::vector<float> get_float_data_from_host_tensor(const HostTensor& tensor) {
+    auto buffer = tensor.get_host_buffer();
+    auto data_span = buffer.view_as<const float>();
+    return std::vector<float>(data_span.begin(), data_span.end());
+}
+
+TEST_F(TransferToDeviceTest, BasicRoundTrip) {
     // Create host tensor with some data
     const Shape shape{1, 1, 32, 32};
     const size_t num_elements = shape.volume();
@@ -41,15 +59,13 @@ TEST_F(TransferToDeviceTest, BasicTransfer) {
 
     TensorSpec tensor_spec(shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{}));
 
-    // Create host tensor using HostBuffer directly
-    std::vector<float> data_copy(host_data);
-    HostBuffer host_buffer{std::move(data_copy)};
-    HostTensor host_tensor{std::move(host_buffer), tensor_spec, TensorTopology{}};
+    // Create host tensor using from_vector
+    auto host_tensor = HostTensor::from_vector(std::vector<float>(host_data), tensor_spec);
 
     // Create device tensor spec with DRAM memory config and allocate on device
     TensorSpec device_tensor_spec(
         shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{BufferType::DRAM}));
-    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, mesh_device_.get());
+    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, *mesh_device_);
 
     ASSERT_TRUE(device_tensor.is_allocated());
 
@@ -57,15 +73,17 @@ TEST_F(TransferToDeviceTest, BasicTransfer) {
     auto& cq = mesh_device_->mesh_command_queue();
     TransferToDevice(cq, host_tensor, device_tensor, /*blocking=*/true);
 
-    // Read data back and verify
-    std::vector<float> result_data(num_elements);
-    distributed::ReadShard(cq, result_data, device_tensor.mesh_buffer(), distributed::MeshCoordinate{0, 0});
+    // Create receiving host tensor and transfer back from device
+    auto result_tensor = create_receiving_host_tensor(device_tensor_spec);
+    TransferToHost(cq, device_tensor, result_tensor, /*blocking=*/true);
 
+    // Compare round-tripped data with original
+    auto result_data = get_float_data_from_host_tensor(result_tensor);
     EXPECT_EQ(result_data, host_data);
 }
 
-// Test that DeviceTensor's topology is updated after transfer
-TEST_F(TransferToDeviceTest, TopologyUpdatedAfterTransfer) {
+// Test that topology is preserved through round-trip transfer
+TEST_F(TransferToDeviceTest, TopologyPreservedAfterRoundTrip) {
     const Shape shape{1, 1, 32, 32};
     const size_t num_elements = shape.volume();
 
@@ -74,13 +92,11 @@ TEST_F(TransferToDeviceTest, TopologyUpdatedAfterTransfer) {
 
     TensorSpec tensor_spec(shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{}));
 
-    std::vector<float> data_copy(host_data);
-    HostBuffer host_buffer{std::move(data_copy)};
-    HostTensor host_tensor{std::move(host_buffer), tensor_spec, TensorTopology{}};
+    auto host_tensor = HostTensor::from_vector(std::vector<float>(host_data), tensor_spec);
 
     TensorSpec device_tensor_spec(
         shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{BufferType::DRAM}));
-    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, mesh_device_.get());
+    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, *mesh_device_);
 
     auto& cq = mesh_device_->mesh_command_queue();
     TransferToDevice(cq, host_tensor, device_tensor, /*blocking=*/true);
@@ -89,16 +105,53 @@ TEST_F(TransferToDeviceTest, TopologyUpdatedAfterTransfer) {
     EXPECT_TRUE(device_tensor.is_allocated());
     EXPECT_EQ(device_tensor.logical_shape(), shape);
     EXPECT_EQ(device_tensor.dtype(), DataType::FLOAT32);
+
+    // Transfer back to host and verify data integrity
+    auto result_tensor = create_receiving_host_tensor(device_tensor_spec);
+    TransferToHost(cq, device_tensor, result_tensor, /*blocking=*/true);
+
+    // Verify the received HostTensor has correct properties
+    EXPECT_EQ(result_tensor.logical_shape(), shape);
+    EXPECT_EQ(result_tensor.dtype(), DataType::FLOAT32);
+
+    // Verify data is correct
+    auto result_data = get_float_data_from_host_tensor(result_tensor);
+    EXPECT_EQ(result_data, host_data);
 }
 
-// Test replication: 1x1 host tensor replicated to full mesh device
-// This test validates replication behavior when the mesh has more than one device.
-TEST_F(TransferToDeviceTest, ReplicationToMultiDeviceMesh) {
-    const auto mesh_shape = mesh_device_->shape();
-    if (mesh_shape.mesh_size() < 2) {
-        GTEST_SKIP() << "Need at least 2 devices for replication test";
+// Test round-trip with different data values to ensure data integrity
+TEST_F(TransferToDeviceTest, DataIntegrityRoundTrip) {
+    const Shape shape{1, 1, 32, 32};
+    const size_t num_elements = shape.volume();
+
+    // Use non-sequential data to better test data integrity
+    std::vector<float> host_data(num_elements);
+    for (size_t i = 0; i < num_elements; ++i) {
+        host_data[i] = static_cast<float>(i * 3.14159f + 1.0f);
     }
 
+    TensorSpec tensor_spec(shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{}));
+
+    auto host_tensor = HostTensor::from_vector(std::vector<float>(host_data), tensor_spec);
+
+    TensorSpec device_tensor_spec(
+        shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{BufferType::DRAM}));
+    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, *mesh_device_);
+
+    ASSERT_TRUE(device_tensor.is_allocated());
+
+    auto& cq = mesh_device_->mesh_command_queue();
+    TransferToDevice(cq, host_tensor, device_tensor, /*blocking=*/true);
+
+    auto result_tensor = create_receiving_host_tensor(device_tensor_spec);
+    TransferToHost(cq, device_tensor, result_tensor, /*blocking=*/true);
+
+    auto result_data = get_float_data_from_host_tensor(result_tensor);
+    EXPECT_EQ(result_data, host_data);
+}
+
+// Test round-trip with L1 buffer type
+TEST_F(TransferToDeviceTest, L1BufferRoundTrip) {
     const Shape shape{1, 1, 32, 32};
     const size_t num_elements = shape.volume();
 
@@ -107,30 +160,23 @@ TEST_F(TransferToDeviceTest, ReplicationToMultiDeviceMesh) {
 
     TensorSpec tensor_spec(shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{}));
 
-    // Create host tensor on 1x1 mesh (single shard)
-    std::vector<float> data_copy(host_data);
-    HostBuffer host_buffer{std::move(data_copy)};
-    HostTensor host_tensor{std::move(host_buffer), tensor_spec, TensorTopology{}};
+    auto host_tensor = HostTensor::from_vector(std::vector<float>(host_data), tensor_spec);
 
-    // Create device tensor on the full mesh using allocate_on_device
+    // Use L1 buffer type instead of DRAM
     TensorSpec device_tensor_spec(
-        shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{BufferType::DRAM}));
-    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, mesh_device_.get());
+        shape, TensorLayout(DataType::FLOAT32, PageConfig(Layout::ROW_MAJOR), MemoryConfig{BufferType::L1}));
+    auto device_tensor = DeviceTensor::allocate_on_device(device_tensor_spec, *mesh_device_);
 
     ASSERT_TRUE(device_tensor.is_allocated());
 
-    // Transfer data - this should replicate to all devices
     auto& cq = mesh_device_->mesh_command_queue();
     TransferToDevice(cq, host_tensor, device_tensor, /*blocking=*/true);
 
-    // Read data back from all devices and verify replication
-    for (size_t row = 0; row < mesh_shape[0]; ++row) {
-        for (size_t col = 0; col < mesh_shape[1]; ++col) {
-            std::vector<float> result_data(num_elements);
-            distributed::ReadShard(cq, result_data, device_tensor.mesh_buffer(), distributed::MeshCoordinate{row, col});
-            EXPECT_EQ(result_data, host_data) << "Data mismatch at coord (" << row << ", " << col << ")";
-        }
-    }
+    auto result_tensor = create_receiving_host_tensor(device_tensor_spec);
+    TransferToHost(cq, device_tensor, result_tensor, /*blocking=*/true);
+
+    auto result_data = get_float_data_from_host_tensor(result_tensor);
+    EXPECT_EQ(result_data, host_data);
 }
 
 }  // namespace
